+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [2.0.8] - 2022-11-15
   - `algoritmia/schemes/bt_scheme.py`. Añadido State como sinónimo de Any.
   - `algoritmia/algorithms/traversers.py`. Cambios en dos comentarios.
+  - `algoritmia/schemes/dac_scheme.py`. Cambiado Iterable por Iterator.
 
 ## [2.0.7] - 2022-10-06
   - `algoritmia/schemes/bt_scheme.py`. El método abstracto `successors()` de la clase `DecisionsSequence` ahora devuelve